+++ Chimera/NiawgStructures.h
@@ -46,9 +46,9 @@ struct channelWave
 	int initType;
 	// variables for dealing with varied waveforms. These only get set when a varied waveform is used, and they serve the purpose of 
 	// carrying relevant info to the end of the program, when these varried waveforms are compiled.
-	int varNum;
+	UINT varNum;
 	std::vector<std::string> varNames;
-	std::vector<int> varTypes;
+	std::vector<long> varTypes;
 	// This should usually just be a single char, but if it's wrong I can use the whole string to debug.
 	std::string delim;
 	// the actual waveform data.
